Refactored I2C device; implemented in emulator which allows child devices to be connected to I2C bus.

+++ code/HAL/I2C.h
@@ -12,4 +12,4 @@
 
 EXTERN_C int32_t hal_i2c_write(uint8_t deviceAddr, uint8_t controlAddr, uint8_t controlData);
 
-EXTERN_C int32_t hal_i2c_read(uint8_t deviceAddr, uint8_t controlAddr, uint8_t* outControlData, int32_t nbytes);
+EXTERN_C int32_t hal_i2c_read(uint8_t deviceAddr, uint8_t controlAddr, uint8_t* outControlData, uint8_t nbytes);
